+++ lib/src/average_hash.dart
@@ -1,6 +1,8 @@
 import 'package:image/image.dart';
-import 'hash_base.dart';
 import 'package:collection/collection.dart';
+
+import 'image_hasher.dart';
+import 'image_utils.dart';
 
 /// Computes the average hash (aHash) of an image.
 ///